Fixed a bug
Fixed a small bug in the random color button click event.  If the text box was blank and then the user picked random color to begin with the colors didn't function properly.  Fixed by fixing the userValue and other small things and giving default values.  Also fixed the clearGrid function to where the squares get deleted, which obviously saves on memory.